# backend/booking-orchestrator/src/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Booking Orchestrator API",
    "description": "Main orchestration service that coordinates authentication, bookings, and course management",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/auth/complete": {
      "get": {
        "operationId": "completeLogin",
        "tags": ["authentication"],
        "summary": "Complete OAuth login",
        "description": "Finalizes the OAuth login process after the adapter redirects back. Sets access and refresh token cookies and redirects to frontend.",
        "parameters": [
          {
            "in": "query",
            "name": "nonce",
            "required": true,
            "description": "Unique nonce from the OAuth flow",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Redirect to frontend with authentication cookies set",
            "headers": {
              "Location": {
                "schema": {
                  "type": "string"
                },
                "description": "Frontend redirect URL"
              },
              "Set-Cookie": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "description": "access_token and refresh_token cookies"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
...
    "/api/v1/auth/refresh": {
      "post": {
        "operationId": "refreshAccessToken",
        "tags": ["authentication"],
        "summary": "Refresh access token",
        "description": "Uses the refresh token cookie to issue new access and refresh tokens.",
        "responses": {
          "204": {
            "description": "Tokens refreshed successfully",
            "headers": {
              "Set-Cookie": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "description": "New access_token and refresh_token cookies"
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
    "/api/v1/auth/internal/assert": {
      "post": {
        "operationId": "oauthAssert",
        "tags": ["authentication"],
        "summary": "Internal OAuth assertion endpoint",
        "description": "Internal endpoint used by OAuth adapters to submit verified user assertions. Not intended for direct client use.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "provider": {
                    "type": "string"
                  },
                  "sub": {
                    "type": "string"
                  },
                  "iat": {
                    "type": "integer"
                  },
                  "nonce": {
                    "type": "string"
                  },
                  "signature": {
                    "type": "string"
                  },
                  "picture": {
                    "type": "string"
                  },
                  "frontendRedirectUrl": {
                    "type": "string"
                  }
                },
                "required": ["provider", "sub", "nonce", "signature"]
              }
            }
          }
        },
        "responses": {
          "204": {
            "$ref": "#/components/responses/NoContent"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
    "/api/v1/courses/{id}": {
      "get": {
        "operationId": "getCourseById",
        "tags": ["courses"],
        "summary": "Get course by ID",
        "description": "Returns detailed information about a specific course.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Course ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Course retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "course": {
                          "$ref": "#/components/schemas/Course"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/api/v1/bookings": {
      "get": {
        "operationId": "getUserBookings",
        "tags": ["bookings"],
        "summary": "Get user bookings",
        "description": "Returns all bookings for the authenticated user.",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Bookings retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "bookings": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Booking"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "502": {
            "$ref": "#/components/responses/BadGateway"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          },
          "504": {
            "$ref": "#/components/responses/GatewayTimeout"
          }
        }
      }
    },
...
    "/api/v1/bookings/reservations/{reservationId}/cancel": {
      "patch": {
        "operationId": "cancelReservation",
        "tags": ["bookings"],
        "summary": "Cancel reservation",
        "description": "Cancels a pending reservation for the authenticated user.",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "reservationId",
            "required": true,
            "description": "Reservation ID to cancel",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "$ref": "#/components/responses/NoContent"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "502": {
            "$ref": "#/components/responses/BadGateway"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          },
          "504": {
            "$ref": "#/components/responses/GatewayTimeout"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access_token"
      }
    },
    "responses": {
      "BadRequest": {
        "description": "The server could not understand the request due to invalid syntax",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Authentication is required and has failed or not been provided",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "Forbidden": {
        "description": "The client does not have access rights to the content",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "NotFound": {
        "description": "The server cannot find the requested resource",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "Conflict": {
        "description": "The request could not be completed due to a conflict",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "InternalServerError": {
        "description": "The server encountered an unexpected condition",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "BadGateway": {
        "description": "The server received an invalid response from the upstream server",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "ServiceUnavailable": {
        "description": "The server is currently unavailable",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "GatewayTimeout": {
        "description": "The server did not receive a timely response from the upstream server",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "NoContent": {
        "description": "The request was successful, but there is no content to return"
      }
    },
    "schemas": {
      "Course": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "slots": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "_id": {
                  "type": "string"
                },
                "start": {
                  "type": "string",
                  "format": "date-time"
                },
                "end": {
                  "type": "string",
                  "format": "date-time"
                },
                "capacity": {
                  "type": "integer"
                },
                "available": {
                  "type": "integer"
                }
              }
            }
          },
          "open": {
            "type": "boolean"
          },
          "priceOptions": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "_id": {
                  "type": "string"
                },
                "numberSlots": {
                  "type": "integer"
                },
                "price": {
                  "type": "string"
                }
              }
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Booking": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "course": {
            "type": "string"
          },
          "user": {
            "type": "string"
          },
          "reservation": {
            "type": "string"
          },
          "slots": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "status": {
            "type": "string",
            "enum": ["confirmed", "cancelled"]
          },
          "transactionId": {
            "type": "string"
          },
          "price": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
...
}